requirements.txt
python-vlc
PySide6
mutagen
pytest